1.9 - Hardware Security
With this update, the system can experience physical damage that
prevents momentary communication with the sensors and even turn them
off, the system will find out if the sensor is already available for use
and will make your calibration automatically.  This is very important,
since the acceleration that the rocket suffers can disconnect the sensor
for a moment.