1.16 - The Preprocessing Review
Thermometer object added.
The file on the SD card always wrote its name in its header, no longer
does.
New function that returns the SD card file name. This is to transmit the
name of the file used.
All pre-processing were reviewed and some parts of the program that
should be conditioned to they spent actually be conditional on their
part.
A counter systems was added to the beginning of the program together
with the RBF directive, and such data is transmitted if the system has
telemetry.
The RBF system was completely revised, the buzzer was tidy, the beeps
number is equivalent to the number of operative systems.
New GPS information is collected and transmitted. A new organization for
the GPS data was made.
There was a reorganization of the data displayed in serial, so that only
what is active is displayed. This system only serves for data analysis,
for in flight is useless.