2.2 - Much News
The Moving Average class has been absolutely rethought, it is now the
fastest moving average filter that could be.
A new template class with a specially crafted data structure for the new
filter, the secret of the speed.
Passing data by reference in some functions.
Some other minor changes.
A new ascii art for a new logo.